+++ crates/orbit-dashboard/assets/dashboard/router.js
@@ -268,3 +268,8 @@ export function initTabs() {
   const ctx = getCtx();
   return initTabsImpl(ctx);
 }
+
+export function setRunDetailSubtab(name) {
+  const ctx = getCtx();
+  setRunDetailSubtabImpl(ctx, name);
+}
