+++ crates/orbit-dashboard/src/lib.rs
@@ -36,6 +36,7 @@ const LOG_TAIL_JS: &str = include_str!("../assets/dashboard/log-tail.js");
 const DIAGNOSTICS_JS: &str = include_str!("../assets/dashboard/diagnostics.js");
 const ROUTER_JS: &str = include_str!("../assets/dashboard/router.js");
 const RUNS_JS: &str = include_str!("../assets/dashboard/runs.js");
+const RUN_DETAIL_JS: &str = include_str!("../assets/dashboard/run-detail.js");
 
 /// Arguments for `orbit web serve` (and the library entry point).
 #[derive(Args, Clone)]
@@ -77,6 +78,7 @@ pub fn serve(runtime: &OrbitRuntime, args: ServeArgs) -> Result<(), OrbitError> 
         .route("/static/diagnostics.js", get(serve_diagnostics_js))
         .route("/static/router.js", get(serve_router_js))
         .route("/static/runs.js", get(serve_runs_js))
+        .route("/static/run-detail.js", get(serve_run_detail_js))
         .route("/healthz", get(healthz))
         .nest("/api", api::router())
         .with_state(runtime);
@@ -226,6 +228,17 @@ async fn serve_runs_js() -> Response {
             HeaderValue::from_static("application/javascript; charset=utf-8"),
         )],
         RUNS_JS,
+    )
+        .into_response()
+}
+
+async fn serve_run_detail_js() -> Response {
+    (
+        [(
+            header::CONTENT_TYPE,
+            HeaderValue::from_static("application/javascript; charset=utf-8"),
+        )],
+        RUN_DETAIL_JS,
     )
         .into_response()
 }
